backend/requirements.txt
# Web Framework
fastapi==0.115.0
uvicorn[standard]==0.32.0
python-multipart==0.0.12

# Database
pymongo>=4.9,<5.0
motor>=3.6.0

# Configuration & Validation
python-dotenv==1.0.1
pydantic==2.9.2
pydantic-settings==2.5.2

# Langchain & RAG
langchain>=0.3.0
langchain-community>=0.3.0
langchain-text-splitters>=0.3.0
langchain-openai>=0.2.0
langchain-groq>=0.1.0
langchain-huggingface>=0.1.0

# Vector Database (ChromaDB - Primary)
chromadb==0.5.0
# Vector Database (Qdrant - Alternative)
qdrant-client==1.12.0

# Embeddings
sentence-transformers==3.3.1
openai==1.54.5

# File Processing
PyPDF2==3.0.1
python-magic==0.4.27

# Memory & Evaluation
scikit-learn==1.5.2

# Utilities
httpx==0.27.2
aiofiles==24.1.0

